+++ src/plugins/data/public/search/types.ts
@@ -56,6 +56,7 @@ export interface ISearchSetup {
   __enhance: (enhancements: SearchEnhancements) => void;
   getDefaultSearchInterceptor: () => ISearchInterceptor;
   df: DataFrameService;
+  abortControllerRef: React.MutableRefObject<AbortController | undefined>;
 }
 
 /**
