+++ src/plugins/data/public/ui/types.ts
@@ -12,9 +12,8 @@ import { SuggestionsComponentProps } from './typeahead/suggestions_component';
  * point.
  */
 
-export interface IUiSetup {
-  abortControllerRef: React.MutableRefObject<AbortController | undefined>;
-}
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface IUiSetup {}
 
 /**
  * Data plugin prewired UI components
