+++ src/plugins/query_enhancements/public/query_assist/components/query_assist_bar.tsx
@@ -46,7 +46,10 @@ export const QueryAssistBar: React.FC<QueryAssistInputProps> = (props) => {
   const previousQuestionRef = useRef<string>();
   const { updateQueryState } = useQueryAssist();
   const uiActions = getUiActions();
-  const { generateQuery, loading } = useGenerateQuery(uiActions, props.data.ui.abortControllerRef);
+  const { generateQuery, loading } = useGenerateQuery(
+    uiActions,
+    props.data.search.abortControllerRef
+  );
 
   useEffect(() => {
     const subscription = queryString.getUpdates$().subscribe((query) => {
